Time in Transit enhancement
Update the zip input in the header so that when it changes, a geocoder
request goes out for city and state for the matching zip.

Updated the minicart update function to include code for updating the
zip, city and state classes. This will handle populating the header at
init if a zip is provided.