Context for prod carousels, dropdown slide out
Product carousels now have context so that carousel is built for each
product page brought into the view.
Altered homepage dropdown so that slide out will extend and allow for
second image on sub menu. Still needs to be properly styled and applied
to all dropdowns (is only on first dropdown now).