Prod page vertical carousel timing
Set vertical carousels to scroll every 3 seconds per Vin's request.
Also set initial scroll to happen -1 so to aid in eliminating
Caroufredsel from rendering items scrolled half way initially; the first
scroll eliminates this so if it happens on page load the user should
only experience it as part of the page load.