productListTemplate hiding discontinued items
Changed render format that adds the hanging tags to hide the parent li
in the product list in the event that the tag IS_DISCONTINUED is set.
Infinite scroll will be added so this will work correctly. If infinite
scroll is removed, then the renderformat will have to be removed so the
pagination count doesn't get thrown off.