Prod list 1st sibling image / version
If user:grp_siblings has images the 2nd in the list (product image is
usually first in this client's lists) will be displayed on product in
product list (will not work for elastic results).
If not image is attrib, reviews will move up to fill space (other wise
reviews not displayed).