prodAttribs['user:prod_ship_expavail'] 1 if set / version
Vin confirmed that the test product was set incorrectly, and that
prodAttribs['user:prod_ship_expavail'] set to 1 indicates expedited
shipping IS available, and 0 IS NOT available.
May still need to change expedited message display for other items in
the cart when a single item is not expedite-able.  Once testing product
attrib is correct, testing can occur.