getShipMethodByID split for prod page & cart
moved the edited version of getShipMethodByID into
beachmart_cartEstArrival because it was possible for the index to return
false from the product page. This caused an error. Since the ground
shipping is all that is needed (ship method can't be changed on prod
page, only in cart) the getShipMethodByID was reverted back to only
grabbing ground for the prod page.