Cart transit time STID matching fix / version
A change in the way the stid was getting added to the li (w/ a
seperator, ie ":") caused the items in cart w/ variations to not have
time in transit added.
New function to generate stid on lower element the same way it is in the
response read in showTransitTimes should allow proper comparison.