product page: vert carousels IS_DISCONTINUED
Items w/ is:tag of discontinued are set to hide. The related/accessory
vertical carousels were being shown w/ no content if all items within
them were discontinued.
Set a check to hide the container for the carousels if there are no
products, or if there are products but all of them are discontinued.
Also attempted to check if product was purchasable but the record came
back incorrect (was purchasable and had discontinued tag).
Left w/ just the tag check for now.