Search results view details button fix
Button was in container that call preview in order to allow hover state
on button to be activated on hover of container.
This was causing the view details click to the product page to trigger
the preview click, running both functions.
Custom hover function to change colors on both items allowed button to
be removed from container to solve the propagation issue.