cart backorder time in transit future date
Added extension to work with dates.
Time in transit for backorder items now shows backorder date to ship on,
as well as adjusts arrival date by adding the current time for delivery
to the backorder ship date.
Ground will be used on all items if there is a backorder item in the
cart.
Expedited not available will also be shown on items that are back order
regardless of it's availability for the product.