recent viewed as a cat list / company sideline
Set up company article to use for recently viewed as a product list
similar to category.
Excluded the article from nav build in quickstart.
Fixed hard coded sideline list in company section.
Altered showRecently viewed to accept and argument to indicate whether
to show it and empty message, or wait until content is present to show
it.