prod list: siblings/reviews showing/hiding / version
regular product lists will show a sibling prod list, if there are
reviews as well they will be shown.
If there is a sibling list and no reviews, reviews will not be shown.
If there is no sibling list, and no reviews, reviews will be shown.
The containing element height will be changed based on all of these
variables.