201402: Brands category tabs/lists
Set brands pages to show product list in three tabs, and to hide items
that don't have the best selling and featured attributes.
Edited the templates for "See All" for these sections to use the same
feature.
Still need to hide tabs that are empty as well as add carousels and
disable "See All" links for tabs w/ fewer than the max carousel size.