Dropdown search button hider feature / version
Added a json file which indicates whether to hide the bestseller,
featured, or clearance search buttons contained within each dropdown.
Intention is to have a utility that will run all of the searches and
build the json file based on results so that it can be run periodically
to update the json file.