prod page price modifiers base price / version
If the user:prod_has_price_modifiers attribute is set and imgselect
variations are being used, each option for each variation with a price
modifier will be checked for the highest possible price modifier and the
value(s) will be added to the base price to display a range in the price
field.
In the style of $100 - $150, rather than just $100.