Magic zoom g.body error fix / version
MZ was getting loaded in beachmall_begin when it was loaded at start up.

Missing resources at that time caused the error.
Moved MZ script into a link in head. This previously may have not worked
before (causing the move to the extension in the first place) because it
had an apptimize embed on it; although that is just speculation).