index: dropdown menu functions removed
The basic links that are in place in case the dropdown template doesn't
render had the hover functions to trigger the dropdowns on them and were
causing an error if they were triggered before the extension they are in
was loaded.
Since the extension they are in also loads the template with the
dropdowns, it is safe for them to stay as actions in the template.