Sitemap categories/noindex / version
Categories use the pretty name and so were listed in custom urls. This
made auto generating them unnecessary so it was commented out.
seo:noindex was changed to noindex for auto generation of product urls
as the entire attrib was not being found correctly.
Product urls for sitemap will need product data to be built in custom
fashion using attrib info instead of pid (as is default).