recently viewed fix
If an item was seen before the homepage was loaded, the element to place
the count in for the floating bar was not yet present, and the count was
not updated there. Added a render format to the element to update the
count on first load, previous functionality will take care of subsequent
updates.
Was set for complete, but should have been depart.